refactor: 테스트코드 외부 클래스 의존성 축소
- 기존코드 스트림 활용해서 간결하게 표현

+++ src/main/java/lotto/domain/Lotto.java
@@ -57,8 +57,8 @@ public class Lotto {
 		}
 	}
 
-	public long matchCount(Lotto lotto) {
-		return lottoNumbers.stream()
+	public int matchCount(Lotto lotto) {
+		return (int)lottoNumbers.stream()
 			.filter(lotto.lottoNumbers::contains)
 			.count();
 	}
